+++ backend/app/core/__init__.py
@@ -1,0 +1,2 @@
+from .database import Base, SessionLocal, engine
+
